Move config.ts to env_vars.ts
Since its not actually a path in the app, it shouldn't be in the app
directory either.
