+++ lib/env_vars.ts
